Implement ui for delivery detials and payments pages

+++ molases_app/src/main/java/com/gierza_molases/molases_app/service/BranchService.java
@@ -3,19 +3,24 @@ package com.gierza_molases.molases_app.service;
 import java.util.List;
 
 import com.gierza_molases.molases_app.dao.BranchDao;
+import com.gierza_molases.molases_app.dao.BranchDeliveryDao;
 import com.gierza_molases.molases_app.dao.CustomerDao;
 import com.gierza_molases.molases_app.model.Branch;
-import com.gierza_molases.molases_app.model.BranchCustomerResponse;
 import com.gierza_molases.molases_app.model.Customer;
+import com.gierza_molases.molases_app.model.response.BranchCustomerResponse;
 
 public class BranchService {
 
 	private final BranchDao branchDao;
 	private final CustomerDao customerDao;
 
-	public BranchService(BranchDao branchDao, CustomerDao customerDao) {
+	private final BranchDeliveryDao branchDeliveryDao;
+
+	public BranchService(BranchDao branchDao, CustomerDao customerDao, BranchDeliveryDao branchDeliveryDao) {
 		this.branchDao = branchDao;
 		this.customerDao = customerDao;
+
+		this.branchDeliveryDao = branchDeliveryDao;
 	}
 
 	/**
@@ -32,7 +37,9 @@ public class BranchService {
 	 * Get all branches with search
 	 */
 	public List<Branch> fetchBranchCursor(Long lastSeenBranchId, String customerFilterName, int pageSize) {
-		return branchDao.fetchNextPage(lastSeenBranchId, customerFilterName, pageSize);
+
+		// we add one to pageSize so that we can check if there is more page to fetch
+		return branchDao.fetchNextPage(lastSeenBranchId, customerFilterName, pageSize + 1);
 	}
 
 	/**
@@ -61,6 +68,12 @@ public class BranchService {
 	}
 
 	public void delete(int branchId) {
+
+		// check first for transaction
+		if (branchDeliveryDao.hasDelivery(branchId)) {
+			throw new IllegalStateException("Cannot delete branch because it has deliveries.");
+		}
+
 		branchDao.deleteBranch(branchId);
 	}
 }
